Add Jenkins API to hubot
This is roughly based on hubot-scripts/jenkins.coffee. The difference
here is that rather than having an actual UI to execute jenkins builds
and commands, we simply expose the API part and we let the calling
code taking care of the authorization part. We also use ADFS_LOGIN to
authenticate to jenkins by-passing the reverse proxy.